수정 : DB 제약조건과 컬럼추가 rasbpi 제약조건 수정 user 컬럼 추가 이름 및 나이

+++ src/models/user.js
@@ -33,7 +33,7 @@ class User extends Model {
                     }
                 },
                 age: {
-                    type: DataTypes.NUMBER(3).UNSIGNED,
+                    type: DataTypes.INTEGER(4).UNSIGNED,
                     allowNull: false,
                     validate: {
                         len: [0, 3]
